fix: harden TensorSysY correctness checks
Reject unsupported tensor constructs and unsafe shapes before lowering, keep observable tensor results through O1, tighten CLI and benchmark failure handling, and update README to match the current gated runtime/test matrix.

+++ src/frontend/ast.h
@@ -223,7 +223,7 @@ class PrimaryExpAST : public BaseAST {
     Def def = def_number;
     BaseASTPtr subExp;
     std::string lVal;
-    int32_t number = 0;
+    int64_t number = 0;
     std::string arrayIdent;
     MulVecType expArray;
